requirements.txt
Pillow
pytesseract
opencv-python
numpy
difflib
